Update to validation
Validation was not checking against the second true/false argument passed to the callback function. Calling "callback(null, false)" when your custom validation method failed to validate a user would still return the route requested, as long as the provided JWT was decoded.

This may not be the desired correction, but it is working for me as a band-aid fix until it is incorporated officially.