Set token in request for access from handler
We set the jwt token on successful authentication in the request object.
It is stored in request.auth.token.

This allows access to the token from handler. It can be used, for
example to proxy request to another server using the user's token,
without needing to issue a new one.