verify the token is still valid in a long running request
Implements a new hapi feature: "async verify(auth) - (optional) a method used to verify the authentication credentials provided are still valid (e.g. not expired or revoked after the initial authentication) where:"

Ideally this would also call something on the `options` to check that the user itself is still valid, not just the token, but that probably needs more though and could result in breaking changes - it can be added later, as the need arises. There is an existing `options.verify()` method, which bypasses all JWT verification, and there is an existing `options.validate()`, but it takes a `request` and `h` as params as well as the decoded token.